refactor: abstract API ID generator to prepare for custom generated formats
