+++ includes/settings/js/src/components/CreateContentModel.jsx
@@ -3,7 +3,7 @@ import { useForm } from "react-hook-form";
 import { Link, useHistory } from "react-router-dom";
 import { ModelsContext } from "../ModelsContext";
 import { insertSidebarMenuItem } from "../utils";
-import { useApiIdGenerator } from "./fields/useApiIdGenerator";
+import { useInputGenerator } from "../hooks";
 import { showSuccess } from "../toasts";
 import Icon from "../../../../components/icons";
 import IconPicker from "./IconPicker";
@@ -31,9 +31,11 @@ export default function CreateContentModel() {
 	const [icon, setIcon] = useState(0);
 	const [descriptionCount, setDescriptionCount] = useState(0);
 	const { dispatch } = useContext(ModelsContext);
-	const { setApiIdGeneratorInput, apiIdFieldAttributes } = useApiIdGenerator({
-		apiFieldId: "slug",
-		setValue,
+	const {
+		setInputGeneratorSourceValue,
+		onChangeGeneratedValue,
+	} = useInputGenerator({
+		setGeneratedValue: (value) => setValue("slug", value),
 	});
 
 	function apiCreateModel(data) {
@@ -111,7 +113,7 @@ export default function CreateContentModel() {
 							className="w-100"
 							ref={register({ required: true, maxLength: 50 })}
 							onChange={(e) => {
-								setApiIdGeneratorInput(event.target.value);
+								setInputGeneratorSourceValue(e.target.value);
 								setSingularCount(e.target.value.length);
 							}}
 						/>
@@ -212,7 +214,9 @@ export default function CreateContentModel() {
 							name="slug"
 							className="w-100"
 							ref={register({ required: true, maxLength: 20 })}
-							{...apiIdFieldAttributes}
+							onChange={(e) =>
+								onChangeGeneratedValue(e.target.value)
+							}
 						/>
 						<p className="field-messages">
 							{errors.slug && errors.slug.type === "required" && (
