+++ includes/settings/js/src/components/fields/Form.jsx
@@ -13,7 +13,7 @@ import NumberFields from "./NumberFields";
 import MultipleChoiceFields from "./MultipleChoiceFields";
 import supportedFields from "./supportedFields";
 import { ModelsContext } from "../../ModelsContext";
-import { useApiIdGenerator } from "./useApiIdGenerator";
+import { useInputGenerator } from "../../hooks";
 import { sprintf, __ } from "@wordpress/i18n";
 
 const { apiFetch } = wp;
@@ -51,10 +51,13 @@ function Form({ id, position, type, editing, storedData }) {
 	const model = query.get("id");
 	const ExtraFields = extraFields[type] ?? null;
 	const currentNumberType = watch("numberType");
-	const { setApiIdGeneratorInput, apiIdFieldAttributes } = useApiIdGenerator({
-		setValue,
+	const {
+		setInputGeneratorSourceValue,
+		onChangeGeneratedValue,
+	} = useInputGenerator({
+		sourceValue: storedData?.name,
 		editing,
-		storedData,
+		setGeneratedValue: (value) => setValue("slug", value),
 	});
 	const originalState = useRef(cloneDeep(models[model]["fields"] || {}));
 	const [previousState, setPreviousState] = useState(storedData);
@@ -278,7 +281,7 @@ function Form({ id, position, type, editing, storedData }) {
 							type="text"
 							ref={register({ required: true, maxLength: 50 })}
 							onChange={(e) => {
-								setApiIdGeneratorInput(e.target.value);
+								setInputGeneratorSourceValue(e.target.value);
 								setNameCount(e.target.value.length);
 								clearErrors("slug");
 							}}
@@ -326,7 +329,9 @@ function Form({ id, position, type, editing, storedData }) {
 							type="text"
 							defaultValue={storedData?.slug}
 							ref={register({ required: true, maxLength: 50 })}
-							{...apiIdFieldAttributes}
+							onChange={(e) =>
+								onChangeGeneratedValue(e.target.value)
+							}
 						/>
 						<p className="field-messages">
 							{errors.slug && errors.slug.type === "required" && (
